Gerneralized images and networking features
This commit makes it so that all images are opt out.
If you want any images created for you by default you will
have to add an export prior to running the script.

Additionally, when Neutron is used, a warning is automatically
added to the root users bashrc to inform the user that they will
have to work with network namespaces to get shell access to the
instances.